# perf(sql): prevent sqlglot from extensive deepcopying every time we create a sqlglot object