--- a/ibis/backends/sql/compiler.py
+++ b/ibis/backends/sql/compiler.py
@@ -5,14 +5,12 @@ import calendar
 import itertools
 import math
 import string
-from collections.abc import Iterator, Mapping
+from collections.abc import Mapping
 from functools import partial, reduce
-from itertools import starmap
 from typing import TYPE_CHECKING, Any, Callable, ClassVar
 
 import sqlglot as sg
 import sqlglot.expressions as sge
-import toolz
 from public import public
 
 import ibis.common.exceptions as com
@@ -132,22 +130,10 @@ class ColGen:
         self.table = table
 
     def __getattr__(self, name: str) -> sge.Column:
-        return sg.column(name, table=self.table)
+        return sg.column(name, table=self.table, copy=False)
 
     def __getitem__(self, key: str) -> sge.Column:
-        return sg.column(key, table=self.table)
-
-
-def paren(expr):
-    """Wrap a sqlglot expression in parentheses."""
-    return sge.Paren(this=sge.convert(expr))
-
-
-def parenthesize(op, arg):
-    if isinstance(op, (ops.Binary, ops.Unary)):
-        return paren(arg)
-    # function calls don't need parens
-    return arg
+        return sg.column(key, table=self.table, copy=False)
 
 
 C = ColGen()
@@ -404,11 +390,11 @@ class SQLGlotCompiler(abc.ABC):
         return sge.If(
             this=sge.convert(condition),
             true=sge.convert(true),
-            false=false if false is None else sge.convert(false),
+            false=None if false is None else sge.convert(false),
         )
 
     def cast(self, arg, to: dt.DataType) -> sge.Cast:
-        return sg.cast(sge.convert(arg), to=self.type_mapper.from_ibis(to))
+        return sg.cast(sge.convert(arg), to=self.type_mapper.from_ibis(to), copy=False)
 
     def _prepare_params(self, params):
         result = {}
@@ -465,7 +451,7 @@ class SQLGlotCompiler(abc.ABC):
                 return result.as_(alias, quoted=self.quoted)
             else:
                 try:
-                    return result.subquery(alias)
+                    return result.subquery(alias, copy=False)
                 except AttributeError:
                     return result.as_(alias, quoted=self.quoted)
 
@@ -475,14 +461,16 @@ class SQLGlotCompiler(abc.ABC):
         # get the root node as a sqlglot select statement
         out = results[op]
         if isinstance(out, sge.Table):
-            out = sg.select(STAR).from_(out)
+            out = sg.select(STAR, copy=False).from_(out, copy=False)
         elif isinstance(out, sge.Subquery):
             out = out.this
 
         # add cte definitions to the select statement
         for cte in ctes:
             alias = sg.to_identifier(aliases[cte], quoted=self.quoted)
-            out = out.with_(alias, as_=results[cte].this, dialect=self.dialect)
+            out = out.with_(
+                alias, as_=results[cte].this, dialect=self.dialect, copy=False
+            )
 
         return out
 
@@ -509,7 +497,7 @@ class SQLGlotCompiler(abc.ABC):
         return self.cast(arg, to)
 
     def visit_ScalarSubquery(self, op, *, rel):
-        return rel.this.subquery()
+        return rel.this.subquery(copy=False)
 
     def visit_Alias(self, op, *, arg, name):
         return arg
@@ -687,12 +675,14 @@ class SQLGlotCompiler(abc.ABC):
         return sge.Between(this=arg, low=lower_bound, high=upper_bound)
 
     def visit_Negate(self, op, *, arg):
-        return -paren(arg)
+        return -sge.paren(arg, copy=False)
 
     def visit_Not(self, op, *, arg):
         if isinstance(arg, sge.Filter):
-            return sge.Filter(this=sg.not_(arg.this), expression=arg.expression)
-        return sg.not_(paren(arg))
+            return sge.Filter(
+                this=sg.not_(arg.this, copy=False), expression=arg.expression
+            )
+        return sg.not_(sge.paren(arg, copy=False))
 
     ### Timey McTimeFace
 
@@ -834,7 +824,7 @@ class SQLGlotCompiler(abc.ABC):
         return arg.is_(NULL)
 
     def visit_NotNull(self, op, *, arg):
-        return arg.is_(sg.not_(NULL))
+        return arg.is_(sg.not_(NULL, copy=False))
 
     def visit_InValues(self, op, *, value, options):
         return value.isin(*options)
@@ -1021,7 +1011,9 @@ class SQLGlotCompiler(abc.ABC):
         return sg.to_identifier(op.param)
 
     def visit_RowID(self, op, *, table):
-        return sg.column(op.name, table=table.alias_or_name, quoted=self.quoted)
+        return sg.column(
+            op.name, table=table.alias_or_name, quoted=self.quoted, copy=False
+        )
 
     # TODO(kszucs): this should be renamed to something UDF related
     def __sql_name__(self, op: ops.ScalarUDF | ops.AggUDF) -> str:
@@ -1066,16 +1058,6 @@ class SQLGlotCompiler(abc.ABC):
 
     ## relations
 
-    def _dedup_name(
-        self, key: str, value: sge.Expression
-    ) -> Iterator[sge.Alias | sge.Column]:
-        """Don't alias columns that are already named the same as their alias."""
-        return (
-            value
-            if isinstance(value, sge.Column) and key == value.name
-            else value.as_(key, quoted=self.quoted)
-        )
-
     @staticmethod
     def _gen_valid_name(name: str) -> str:
         """Generate a valid name for a value expression.
@@ -1089,9 +1071,14 @@ class SQLGlotCompiler(abc.ABC):
 
     def _cleanup_names(self, exprs: Mapping[str, sge.Expression]):
         """Compose `_gen_valid_name` and `_dedup_name` to clean up names in projections."""
-        return starmap(
-            self._dedup_name, toolz.keymap(self._gen_valid_name, exprs).items()
-        )
+
+        for name, value in exprs.items():
+            name = self._gen_valid_name(name)
+            if isinstance(value, sge.Column) and name == value.name:
+                # don't alias columns that are already named the same as their alias
+                yield value
+            else:
+                yield value.as_(name, quoted=self.quoted, copy=False)
 
     def visit_Select(self, op, *, parent, selections, predicates, sort_keys):
         # if we've constructed a useless projection return the parent relation
@@ -1101,18 +1088,20 @@ class SQLGlotCompiler(abc.ABC):
         result = parent
 
         if selections:
-            result = sg.select(*self._cleanup_names(selections)).from_(result)
+            result = sg.select(*self._cleanup_names(selections), copy=False).from_(
+                result, copy=False
+            )
 
         if predicates:
-            result = result.where(*predicates)
+            result = result.where(*predicates, copy=False)
 
         if sort_keys:
-            result = result.order_by(*sort_keys)
+            result = result.order_by(*sort_keys, copy=False)
 
         return result
 
     def visit_DummyTable(self, op, *, values):
-        return sg.select(*self._cleanup_names(values))
+        return sg.select(*self._cleanup_names(values), copy=False)
 
     def visit_UnboundTable(
         self, op, *, name: str, schema: sch.Schema, namespace: ops.Namespace
@@ -1143,12 +1132,14 @@ class SQLGlotCompiler(abc.ABC):
         return parent
 
     def visit_JoinChain(self, op, *, first, rest, values):
-        result = sg.select(*self._cleanup_names(values)).from_(first)
+        result = sg.select(*self._cleanup_names(values), copy=False).from_(
+            first, copy=False
+        )
 
         for link in rest:
             if isinstance(link, sge.Alias):
                 link = link.this
-            result = result.join(link)
+            result = result.join(link, copy=False)
         return result
 
     def visit_JoinLink(self, op, *, how, table, predicates):
@@ -1189,18 +1180,18 @@ class SQLGlotCompiler(abc.ABC):
 
     def visit_Aggregate(self, op, *, parent, groups, metrics):
         sel = sg.select(
-            *self._cleanup_names(groups), *self._cleanup_names(metrics)
-        ).from_(parent)
+            *self._cleanup_names(groups), *self._cleanup_names(metrics), copy=False
+        ).from_(parent, copy=False)
 
         if groups:
-            sel = sel.group_by(*self._generate_groups(groups.values()))
+            sel = sel.group_by(*self._generate_groups(groups.values()), copy=False)
 
         return sel
 
     @classmethod
     def _add_parens(cls, op, sg_expr):
         if isinstance(op, cls.NEEDS_PARENS):
-            return paren(sg_expr)
+            return sge.paren(sg_expr, copy=False)
         return sg_expr
 
     def visit_Filter(self, op, *, parent, predicates):
@@ -1209,104 +1200,126 @@ class SQLGlotCompiler(abc.ABC):
             for raw_predicate, predicate in zip(op.predicates, predicates)
         )
         try:
-            return parent.where(*predicates)
+            return parent.where(*predicates, copy=False)
         except AttributeError:
-            return sg.select(STAR).from_(parent).where(*predicates)
+            return (
+                sg.select(STAR, copy=False)
+                .from_(parent, copy=False)
+                .where(*predicates, copy=False)
+            )
 
     def visit_Sort(self, op, *, parent, keys):
         try:
-            return parent.order_by(*keys)
+            return parent.order_by(*keys, copy=False)
         except AttributeError:
-            return sg.select(STAR).from_(parent).order_by(*keys)
+            return (
+                sg.select(STAR, copy=False)
+                .from_(parent, copy=False)
+                .order_by(*keys, copy=False)
+            )
 
     def visit_Union(self, op, *, left, right, distinct):
         if isinstance(left, (sge.Table, sge.Subquery)):
-            left = sg.select(STAR).from_(left)
+            left = sg.select(STAR, copy=False).from_(left, copy=False)
 
         if isinstance(right, (sge.Table, sge.Subquery)):
-            right = sg.select(STAR).from_(right)
+            right = sg.select(STAR, copy=False).from_(right, copy=False)
 
         return sg.union(
             left.args.get("this", left),
             right.args.get("this", right),
             distinct=distinct,
+            copy=False,
         )
 
     def visit_Intersection(self, op, *, left, right, distinct):
         if isinstance(left, (sge.Table, sge.Subquery)):
-            left = sg.select(STAR).from_(left)
+            left = sg.select(STAR, copy=False).from_(left, copy=False)
 
         if isinstance(right, (sge.Table, sge.Subquery)):
-            right = sg.select(STAR).from_(right)
+            right = sg.select(STAR, copy=False).from_(right, copy=False)
 
         return sg.intersect(
             left.args.get("this", left),
             right.args.get("this", right),
             distinct=distinct,
+            copy=False,
         )
 
     def visit_Difference(self, op, *, left, right, distinct):
         if isinstance(left, (sge.Table, sge.Subquery)):
-            left = sg.select(STAR).from_(left)
+            left = sg.select(STAR, copy=False).from_(left, copy=False)
 
         if isinstance(right, (sge.Table, sge.Subquery)):
-            right = sg.select(STAR).from_(right)
+            right = sg.select(STAR, copy=False).from_(right, copy=False)
 
         return sg.except_(
             left.args.get("this", left),
             right.args.get("this", right),
             distinct=distinct,
+            copy=False,
         )
 
     def visit_Limit(self, op, *, parent, n, offset):
         # push limit/offset into subqueries
         if isinstance(parent, sge.Subquery) and parent.this.args.get("limit") is None:
-            result = parent.this
+            result = parent.this.copy()
             alias = parent.alias
         else:
-            result = sg.select(STAR).from_(parent)
+            result = sg.select(STAR, copy=False).from_(parent, copy=False)
             alias = None
 
         if isinstance(n, int):
-            result = result.limit(n)
+            result = result.limit(n, copy=False)
         elif n is not None:
-            result = result.limit(sg.select(n).from_(parent).subquery())
+            result = result.limit(
+                sg.select(n, copy=False).from_(parent, copy=False).subquery(copy=False),
+                copy=False,
+            )
         else:
             assert n is None, n
             if self.no_limit_value is not None:
-                result = result.limit(self.no_limit_value)
+                result = result.limit(self.no_limit_value, copy=False)
 
         assert offset is not None, "offset is None"
 
         if not isinstance(offset, int):
             skip = offset
-            skip = sg.select(skip).from_(parent).subquery()
+            skip = (
+                sg.select(skip, copy=False)
+                .from_(parent, copy=False)
+                .subquery(copy=False)
+            )
         elif not offset:
             if alias is not None:
-                return result.subquery(alias)
+                return result.subquery(alias, copy=False)
             return result
         else:
             skip = offset
 
-        result = result.offset(skip)
+        result = result.offset(skip, copy=False)
         if alias is not None:
-            return result.subquery(alias)
+            return result.subquery(alias, copy=False)
         return result
 
     def visit_Distinct(self, op, *, parent):
-        return sg.select(STAR).distinct().from_(parent)
+        return (
+            sg.select(STAR, copy=False).distinct(copy=False).from_(parent, copy=False)
+        )
 
     def visit_DropNa(self, op, *, parent, how, subset):
         if subset is None:
             subset = [
-                sg.column(name, table=parent.alias_or_name, quoted=self.quoted)
+                sg.column(
+                    name, table=parent.alias_or_name, quoted=self.quoted, copy=False
+                )
                 for name in op.schema.names
             ]
 
         if subset:
             predicate = reduce(
                 sg.and_ if how == "any" else sg.or_,
-                (sg.not_(col.is_(NULL)) for col in subset),
+                (sg.not_(col.is_(NULL), copy=False) for col in subset),
             )
         elif how == "all":
             predicate = FALSE
@@ -1317,9 +1330,13 @@ class SQLGlotCompiler(abc.ABC):
             return parent
 
         try:
-            return parent.where(predicate)
+            return parent.where(predicate, copy=False)
         except AttributeError:
-            return sg.select(STAR).from_(parent).where(predicate)
+            return (
+                sg.select(STAR, copy=False)
+                .from_(parent, copy=False)
+                .where(predicate, copy=False)
+            )
 
     def visit_FillNa(self, op, *, parent, replacements):
         if isinstance(replacements, Mapping):
@@ -1332,26 +1349,33 @@ class SQLGlotCompiler(abc.ABC):
             }
         exprs = {
             col: (
-                self.f.coalesce(sg.column(col, quoted=self.quoted), sge.convert(alt))
+                self.f.coalesce(
+                    sg.column(col, quoted=self.quoted, copy=False),
+                    sge.convert(alt),
+                )
                 if (alt := mapping.get(col)) is not None
                 else sg.column(col, quoted=self.quoted)
             )
             for col in op.schema.keys()
         }
-        return sg.select(*self._cleanup_names(exprs)).from_(parent)
+        return sg.select(*self._cleanup_names(exprs), copy=False).from_(
+            parent, copy=False
+        )
 
     def visit_CTE(self, op, *, parent):
         return sg.table(parent.alias_or_name, quoted=self.quoted)
 
     def visit_View(self, op, *, child, name: str):
         if isinstance(child, sge.Table):
-            child = sg.select(STAR).from_(child)
+            child = sg.select(STAR, copy=False).from_(child, copy=False)
+        else:
+            child = child.copy()
 
         if isinstance(child, sge.Subquery):
             return child.as_(name, quoted=self.quoted)
         else:
             try:
-                return child.subquery(name)
+                return child.subquery(name, copy=False)
             except AttributeError:
                 return child.as_(name, quoted=self.quoted)
 
@@ -1359,7 +1383,7 @@ class SQLGlotCompiler(abc.ABC):
         return sg.parse_one(query, read=self.dialect)
 
     def visit_SQLQueryResult(self, op, *, query, schema, source):
-        return sg.parse_one(query, dialect=self.dialect).subquery()
+        return sg.parse_one(query, dialect=self.dialect).subquery(copy=False)
 
     def visit_JoinTable(self, op, *, parent, index):
         return parent