@cpcloud 2024-03-08
Hm, the failing tests seem to be exposing a bug in sqlglot.