@kszucs 2024-03-12
After `main` bumped the sqlglot version the current benchmark comparison looks like:

```
------------------------------------------------------------------------------ benchmark 'test_compile[large-duckdb]': 2 tests -------------------------------------------------------------------------------
Name (time in ms)                                 Min                Max               Mean            StdDev             Median               IQR            Outliers       OPS            Rounds  Iterations
--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
test_compile[large-duckdb] (0408_bc9d757)     11.5417 (1.24)     13.2755 (1.0)      12.4163 (1.25)     0.4392 (1.0)      12.3671 (1.29)     0.6514 (1.38)         11;0   80.5394 (0.80)         28           1
test_compile[large-duckdb] (NOW)               9.2860 (1.0)      31.8314 (2.40)      9.9179 (1.0)      2.3738 (5.40)      9.5708 (1.0)      0.4714 (1.0)           1;2  100.8283 (1.0)          89           1
--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
```