@kszucs 2024-03-13
Thanks for the mssql workaround! Merging.